Система аутентификации
Аутентификация и проверка прав на разделы администрирования

+++ app/Http/Controllers/UserController.php
@@ -3,13 +3,24 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\User;
 use App\Right;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(Request $request)
     {
+        if(!Auth::user()->rights()->where('name','add_admins')->exists())
+        {
+            return redirect()->route('admin');
+        }
+
         $deleted = $request->has('deleted')? $request->deleted : null;
         $users = User::orderBy('name')->paginate(10);
         return view('admin.users', ['users' => $users, 'deleted' => $deleted]);
@@ -17,17 +28,26 @@ class UserController extends Controller
 
     public function show()
     {
-        $user = User::findOrFail(1); // AUTH!!!
+        $user = Auth::user();
         return view('user.profile', ['user' => $user]);
     }
 
     public function edit(User $user)
     {
+        if(!Auth::user()->rights()->where('name','add_admins')->exists())
+        {
+            return redirect()->back();
+        }
         return view('admin.profile', ['user' => $user]);
     }
 
     public function update(Request $request, User $user)
     {
+        if(!Auth::user()->rights()->where('name','add_admins')->exists())
+        {
+            return redirect()->back();
+        }
+
         $vbRight = Right::where('name','view_bookings')->first();
         $ecRight = Right::where('name','edit_content')->first();
         $aaRight = Right::where('name','add_admins')->first();
@@ -52,7 +72,7 @@ class UserController extends Controller
 
     public function updateAuth(Request $request)
     {
-        $user = User::findOrFail(1); // AUTH!!!
+        $user = Auth::user();
         $user->name = $request->name;
         $user->updated_at = now();
         $user->save();
@@ -62,6 +82,11 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        if(!Auth::user()->rights()->where('name','add_admins')->exists())
+        {
+            return redirect()->back();
+        }
+
         $id = $user->id;
 
         $user->rights()->detach();
@@ -70,10 +95,9 @@ class UserController extends Controller
         return redirect('/admin/users?deleted='.$id);
     }
 
-    public function destroyAuth() // It may be UNREGISTER
+    public function destroyAuth()
     {
-        $user = User::findOrFail(1); // AUTH!!!
-        // LOGOUT first ???
+        $user = Auth::user();
 
         $user->rights()->detach();
         $user->delete();
